Functinal Programming FTW
When we introduced tty-cursor, we had to deal with Threads, Exceptions
and that combinantion lead to an ugly global variable + `at_exit`
workaround to print errors only after the cursor thread stopped.

Using the Result monad helped cleaning this up. I had to add a
`on_failure` callback to Result, but I'll try to upstream it.