Load end_of_life JSON database from dynamic path
When installed on a fresh Ruby installation without the source code
cloned, the JSON file couldn't be found because it was looking at the
cwd of the running process instead of the Gem's lib directory path.

I wish that Rubygems had a way of exposing a Gem's installation
directory programatically, but I couldn't find it quickly. `Gem.datadir`
didn't do what I wanted! So, we'll do it like it's 2014...