Fetch files asynchronously
This uses the `async` gem to parallelize fetching files from the GitHub API.
The app is heavily IO-bound, so this is a good way to speed it up.

In my tests, this decreased the time to print results from 25s to around 2s.

Note: We're not taking advantage of the FiberScheduler here, because we would
need Ruby 3.1 to run async 2.x.