Handle empty/invalid version files
We use `Gem::Version` to store the ruby version for a particular file. `Gem::Version` is pretty forgiving and will accept empty strings as valid versions, parsing them as "0". This commit introduces a catch-all clause to ensure we don't return a version 0, which always takes precedence over any other version when comparing.